+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import { Form } from './Form';
+import { Todo } from './Todo';
 import registerServiceWorker from './registerServiceWorker';
 import 'semantic-ui-css/semantic.min.css';
 
-ReactDOM.render(<Form />, document.getElementById('root'));
+ReactDOM.render(<Todo />, document.getElementById('root'));
 registerServiceWorker();
